+++ lib/lynx_view.dart
@@ -1,5 +1,11 @@
 export 'package:lynx_view_platform_interface/lynx_view_platform_interface.dart'
-    show LynxLoadError, LynxMessage;
+    show
+        LynxInstanceMemory,
+        LynxLoadError,
+        LynxMemoryPressureLevel,
+        LynxMemoryUsage,
+        LynxMessage;
 
+export 'src/lynx_memory.dart' show LynxMemory;
 export 'src/lynx_view.dart';
 export 'src/lynx_view_controller.dart';